On the javahl-ra branch:
Bring up-to-date with JavaHL changes on trunk@1405517


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/javahl-ra@1425692 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/bindings/javahl/native/CreateJ.cpp
@@ -208,7 +208,7 @@ CreateJ::Checksum(const svn_checksum_t *checksum)
 
   jbyteArray jdigest = JNIUtil::makeJByteArray(
                             (const signed char *)checksum->digest,
-                            svn_checksum_size(checksum));
+                            (int)svn_checksum_size(checksum));
   if (JNIUtil::isExceptionThrown())
     POP_AND_RETURN_NULL;
 
@@ -1099,7 +1099,8 @@ jobject CreateJ::PropertyMap(apr_hash_t *prop_hash)
         POP_AND_RETURN_NULL;
 
       jbyteArray jpropVal = JNIUtil::makeJByteArray(
-                                    (const signed char *)val->data, val->len);
+                                    (const signed char *)val->data,
+                                    (int)val->len);
       if (JNIUtil::isJavaExceptionThrown())
         POP_AND_RETURN_NULL;
